MVC code restructure
Restructured code from index.JS into user and channel routes with matching controllers

Converted all requests to axios package

Fixed a bug where imaginary users could be added to a channel

Known issues:
Broke the error message when failing a login. temporarily just redirects to index
