+++ src/data/backups/usersBackup.json
@@ -1,1 +1,1 @@
-{"ass":{"memberID":"ass","phoneNumber":"8043632795","password":"TempPass123!","approvedChannels":"Welcome"}}
+{"Test Testerman":{"memberID":"Test Testerman","chatToken":"eyJ0eXAiOiJDUlQiLCJhbGciOiJIUzUxMiJ9.eyJpYXQiOjE2NTQwMjYzNDUsImp0aSI6ImRmMzVmNDBhLTFhOGEtNDNhYi1hNGM1LTk4YTkxMzU5ZGM1ZCIsInN1YiI6IjQwNDZkMzBhLTMxMGItNGVlOS05MGVlLTk3YjZkYzRmOGUyNyIsIm0iOiJUZXN0IFRlc3Rlcm1hbiIsInQiOjQzMjAwLCJjaGxzIjoie1wiMFwiOntcInJlYWRcIjp0cnVlLFwid3JpdGVcIjp0cnVlfX0iLCJzdHQiOiJ7fSJ9.4mmdko9AzR1mayh0P8Gf4RSV62d7nuXdpcaDCPWMkgI56lrcmjJN8aLibYugvX7U1mCHNJYDleH4vq6NioEJrw","phoneNumber":"8043632795","password":"Test Testerman","approvedChannels":["Welcome"]}}
