+++ src/data/channels.json
@@ -1,1 +1,1 @@
-{"test":{"members":["test"],"phoneNumber":"","creationDate":"5/23/2022"},"Welcome":{"members":["test","ass"],"phoneNumber":"","creationDate":"5/23/2022"}}
+{"test":{"members":[""],"phoneNumber":"","creationDate":"5/31/2022"},"Welcome":{"members":["Test Testerman"],"phoneNumber":"","creationDate":"5/31/2022"}}
