+++ src/data/users.json
@@ -1,1 +1,1 @@
-{"ass":{"memberID":"ass","chatToken":"","phoneNumber":"1234567890","password":"return token","approvedChannels":["Welcome"]}}
+{"Test Testerman":{"memberID":"Test Testerman","chatToken":"eyJ0eXAiOiJDUlQiLCJhbGciOiJIUzUxMiJ9.eyJpYXQiOjE2NTQwMjYzNDgsImp0aSI6IjM0NTZlZGNlLWI2YmMtNGU3NC1hNWJkLWMxYTVjZGQ2YWI3NyIsInN1YiI6IjQwNDZkMzBhLTMxMGItNGVlOS05MGVlLTk3YjZkYzRmOGUyNyIsIm0iOiJUZXN0IFRlc3Rlcm1hbiIsInQiOjQzMjAwLCJjaGxzIjoie1wiMFwiOntcInJlYWRcIjp0cnVlLFwid3JpdGVcIjp0cnVlfX0iLCJzdHQiOiJ7fSJ9.IWGOeFWfwsCarhpifsCSuadjSoMJ3fszpVhIL3uj9fXIAlFApkvcB5rYym4ui297WXWha-7LLq-D4Hd4u6T-Jg","phoneNumber":"8043632795","password":"Test Testerman","approvedChannels":["Welcome"]}}
