+++ src/routes/Users.js
@@ -4,6 +4,7 @@ const router = express.Router()
 const controller = require("../controllers/UserController");
 
 router.get('/', controller.view)
-router.post('/sign-up', controller.signUp)
+router.post('/sign_up', controller.signUp)
+router.post('/sign_in',controller.signIn)
 
 module.exports = router
