+++ src/views/chRouteTest.html
@@ -26,12 +26,12 @@
                 </label>
             </div>
             <div class="form-row-last">
-                <input formaction="/chAddUser" formmethod="post" type="submit" name="AddUser" class="adduser" value="Add User">
-                <input formaction="/chPopUser" formmethod="post" type="submit" name="PopUser" class="PopUser" value="Remove User">
-                <input formaction="/chCreateChannel" formmethod="post" type="submit" name="CreateChannel" class="CreateChannel" value="Create New Channel">
-                <input formaction="/chListUsers" formmethod="post" type="submit" name="ListUsers" class="ListUsers" value="List Users">
-                <input formaction="/update_token" formmethod="post" type="submit" name="UpdateToken" class="UpdateToken" value="Update Token">
-                <input formaction="/users/adduser" formmethod="post" type="submit" name="AddUser" class="AddUser" value="AddUser">
+                <input formaction="/channel/add" formmethod="post" type="submit" name="AddUser" class="adduser" value="Add User">
+                <input formaction="/channel/pop" formmethod="post" type="submit" name="PopUser" class="PopUser" value="Remove User">
+                <input formaction="/channel/new" formmethod="post" type="submit" name="CreateChannel" class="CreateChannel" value="Create New Channel">
+                <input formaction="/channel/list" formmethod="post" type="submit" name="ListUsers" class="ListUsers" value="List Users">
+                <input formaction="/channel/update_token" formmethod="post" type="submit" name="UpdateToken" class="UpdateToken" value="Update Token">
+                <input formaction="/users/sign_up" formmethod="post" type="submit" name="AddUser" class="AddUser" value="AddUser">
 
             </div>
         </div>
