+++ src/views/index.html
@@ -60,7 +60,7 @@
 							</label>
 						</div>
 						<div class="form-row-last" >
-							<input formaction="/sign_up" formmethod="post" type="submit" name="register" class="register" value="Register">
+							<input formaction="/users/sign_up" formmethod="post" type="submit" name="register" class="register" value="Register">
 						</div>
 					</div>
 				</form>
@@ -81,7 +81,7 @@
 							</label>
 						</div>
 						<div class="form-row-last">
-							<input formaction="/sign_in" formmethod="post" type="submit" name="register" class="register" value="Sign In">
+							<input formaction="/users/sign_in" formmethod="post" type="submit" name="register" class="register" value="Sign In">
 						</div>
 					</div>
 				</form>
